Move ExperimentalPruner to a separate file
